fix: Start moving adapters to shared base file to allow for updates
